test(ui): move backlink tests to features/backlinks.test.ts
Co-locate backlink test suites (parseBacklinks, backlinkField,
backlinkEditFilter, Enter keymap) with features/backlinks.ts, leaving
only the CTCodeEditor component suite in ct-code-editor.test.ts.
